delete revs

+++ src/database/models/pat_item.rs
@@ -237,15 +237,15 @@ impl PersonalAccessToken {
             return Ok(());
         }
 
-        for (id, token, user_id) in clear_pats {
-            redis
-                .delete_many([
+        redis
+            .delete_many(clear_pats.into_iter().flat_map(|(id, token, user_id)| {
+                [
                     (PATS_NAMESPACE, id.map(|i| i.0.to_string())),
                     (PATS_TOKENS_NAMESPACE, token),
                     (PATS_USERS_NAMESPACE, user_id.map(|i| i.0.to_string())),
-                ])
-                .await?;
-        }
+                ]
+            }))
+            .await?;
 
         Ok(())
     }
